MDEV-24667 LOAD DATA INFILE/inserted rows not written to binlog
Problem:- In master slave replication , when master is using statement format
and slave is using mixed format, sometime we can have binlog which is not
written. For example LOAD DATA INFILE is a unsafe statement , it is logged as
a row format, and if it is on temporary table , it wont be logged. If the
next statement is copy data from the tmp table to normal table, it is logged
as a statement format on the slave , because it thinks this statement is
unsafe. As we can see we did not log LOAD DATA INFILE and we simply logging
insert into from * into statement format , we have effectively lost the
tmp table data.

Solution:- On master in mixed format if tmp table is created thd->temporary_tables
is set to not null. This will make THD::reset_for_next_command to not change
the binlog format back to stmt. So till the tmp table exist all the logs will
be written in row format. On master THD::has_thd_temporary_tables() is used
to determine whether we have temporary tables or not. But this does not work
for the slave , slave shifts temporary tables from thd->temporary_tables to
rgi_slave->rli->save_temporary_tables, so this will make has_thd_temporary_tables()
check to return false on slave. So instead of this function I am calling
has_temporary_tables() , which also checks slave thread tmp tables.