MDEV-26713 UTF8 support on Windows, convert my.ini from ANSI to UTF8
Handle upgrade - on Windows that is capable of UTF8 codepage, convert
entries in my.ini from ANSI to UTF8, during upgrade

Reason is that for the server, paths such as datadir, innodb directories,
location of SSL certificates must now be utf8.

For the client programs , user name,  database etc should be in UTF8,
too, as UTF-8 is now the default charset.