MDEV-27349 ASAN use-after-poison in Query_log_event
The poisoning comaplaint was caused by that
Rollback ALTER does not assign anything to its THD::lex::sql_command,
so it found the value from a previous event and 'cos it was not
ALTER, an inmpropriate branch of ctor was attempted to be
executed.

Fixed with extending `wrte_bin_log` logger function to accept
a `direct` argument that is set to `true`. This value of `true`
makes ctor uninterested in `sql_command` and it's the only value
that ALTER can be logged with that is always directly bypassing
caches.